refactor: centralize IconMapper logic
Centralize all the logic of the Icon Mapper in a single file in the design system module. Centralized the drawable icons too, this deletes duplicated code and drawables. For future updates when adding more icons this eases the process.
