+++ core/designsystem/src/main/res/drawable/assignment_24px.xml
