+++ core/designsystem/src/main/res/drawable/asterisk_24px.xml
