+++ core/designsystem/src/main/res/drawable/bedtime_24px.xml
