+++ core/designsystem/src/main/res/drawable/block_24px.xml
