+++ core/designsystem/src/main/res/drawable/book_ribbon_24px.xml
