+++ core/designsystem/src/main/res/drawable/cleaning_24px.xml
