+++ core/designsystem/src/main/res/drawable/dentistry_24px.xml
