+++ core/designsystem/src/main/res/drawable/directions_bike_24px.xml
