+++ core/designsystem/src/main/res/drawable/directions_run_24px.xml
