+++ core/designsystem/src/main/res/drawable/edit_24px.xml
