+++ core/designsystem/src/main/res/drawable/fitness_center_24px.xml
