+++ core/designsystem/src/main/res/drawable/language_chinese_dayi_24px.xml
