+++ core/designsystem/src/main/res/drawable/local_laundry_service_24px.xml
