+++ core/designsystem/src/main/res/drawable/nutrition_24px.xml
