+++ core/designsystem/src/main/res/drawable/pets_24px.xml
