+++ core/designsystem/src/main/res/drawable/recycling_24px.xml
