+++ core/designsystem/src/main/res/drawable/remove_24px.xml
