+++ core/designsystem/src/main/res/drawable/self_improvement_24px.xml
