+++ core/designsystem/src/main/res/drawable/shopping_basket_24px.xml
