+++ core/designsystem/src/main/res/drawable/shower_24px.xml
