+++ core/designsystem/src/main/res/drawable/washoku_24px.xml
