+++ core/designsystem/src/main/res/drawable/water_drop_24px.xml
