+++ core/widgets/build.gradle.kts
@@ -43,6 +43,7 @@ android {
 
 dependencies {
     implementation(project(":core:domain"))
+    implementation(project(":core:designsystem"))
 
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.compose.foundation)
