--- feature/habits/src/main/res/drawable/assignment_24px.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal"
-    android:autoMirrored="true">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M202.87,848.13Q165.33,848.13 138.6,821.4Q111.87,794.67 111.87,757.13L111.87,202.87Q111.87,165.33 138.6,138.6Q165.33,111.87 202.87,111.87L362.98,111.87Q377.65,75.39 409.35,53.63Q441.04,31.87 480,31.87Q518.96,31.87 550.65,53.63Q582.35,75.39 597.02,111.87L757.13,111.87Q794.67,111.87 821.4,138.6Q848.13,165.33 848.13,202.87L848.13,757.13Q848.13,794.67 821.4,821.4Q794.67,848.13 757.13,848.13L202.87,848.13ZM282.87,677.13L559.28,677.13L559.28,597.13L282.87,597.13L282.87,677.13ZM282.87,520L677.13,520L677.13,440L282.87,440L282.87,520ZM282.87,362.87L677.13,362.87L677.13,282.87L282.87,282.87L282.87,362.87ZM502.82,162.82Q511.91,153.72 511.91,140Q511.91,126.28 502.82,117.18Q493.72,108.09 480,108.09Q466.28,108.09 457.18,117.18Q448.09,126.28 448.09,140Q448.09,153.72 457.18,162.82Q466.28,171.91 480,171.91Q493.72,171.91 502.82,162.82Z"/>
-</vector>
