--- feature/habits/src/main/res/drawable/asterisk_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M434.5,848.13L434.5,589.15L251.85,772.57L187.2,707.91L370.13,525.74L111.87,525.74L111.87,434.74L370.85,434.74L187.43,252.09L251.85,187.2L434.5,370.61L434.5,111.87L525.5,111.87L525.5,370.13L707.67,187.2L772.57,252.09L589.15,434.74L848.13,434.74L848.13,525.74L589.63,525.74L772.57,707.91L707.91,772.57L525.5,589.39L525.5,848.13L434.5,848.13Z"/>
-</vector>
