--- feature/habits/src/main/res/drawable/block_24px.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal"
-    android:autoMirrored="true">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M320.89,855.91Q246.46,823.7 191.26,768.5Q136.07,713.3 103.97,638.78Q71.87,564.26 71.87,479.52Q71.87,394.93 104.01,320.54Q136.15,246.14 191.24,191.11Q246.34,136.07 320.82,103.97Q395.3,71.87 479.99,71.87Q564.67,71.87 639.16,103.97Q713.65,136.07 768.75,191.11Q823.85,246.14 855.99,320.54Q888.13,394.93 888.13,479.52Q888.13,564.26 856.03,638.78Q823.93,713.3 768.74,768.5Q713.54,823.7 639.09,855.91Q564.63,888.13 479.98,888.13Q395.33,888.13 320.89,855.91ZM480,797.13Q532.09,797.13 580.65,780.83Q629.22,764.52 669.78,732.72L226.8,289.98Q195.48,331.02 179.17,379.35Q162.87,427.67 162.87,479.52Q162.87,612.33 255.15,704.73Q347.43,797.13 480,797.13ZM733.43,669.07Q764.52,628.02 780.83,579.7Q797.13,531.37 797.13,479.52Q797.13,346.92 704.85,254.9Q612.57,162.87 480,162.87Q428.15,162.87 379.95,178.93Q331.74,195 290.7,226.09L733.43,669.07ZM480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Q480,479.52 480,479.52Z"/>
-</vector>
