--- feature/habits/src/main/res/drawable/book_ribbon_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M480.72,817.7Q432.48,779.94 376.48,757.98Q320.48,736.02 260,736.02Q218.72,736.02 179.05,747.14Q139.39,758.26 102.87,777.78Q78.52,790.46 55.55,776.66Q32.59,762.87 32.59,736.28L32.59,254.52Q32.59,240.17 39.4,227.54Q46.22,214.91 59.85,208.48Q106.33,184.72 156.57,173.08Q206.8,161.43 259.28,161.43Q332.8,161.43 384.09,178.32Q435.37,195.2 493.93,229.48Q506.61,236.67 513.3,247.18Q520,257.7 520,272.61L520,690.85Q564,669.85 608.5,659.35Q653,648.85 700,648.85Q736,648.85 770.5,654.85Q805,660.85 840,672.85Q840,672.85 840,672.85Q840,672.85 840,672.85L840,183.48Q855.48,188.48 870.46,194.48Q885.43,200.48 900.15,208.48Q913.78,214.91 920.6,227.54Q927.41,240.17 927.41,254.52L927.41,739.87Q927.41,765.5 904.57,778.1Q881.72,790.7 857.13,777.78Q820.61,758.26 780.95,747.14Q741.28,736.02 700,736.02Q640,736.02 584.36,758.1Q528.72,780.17 480.72,817.7ZM620,561.67L620,113.54L740,71.87L740,519.76L620,561.67Z"/>
-</vector>
