--- feature/habits/src/main/res/drawable/cleaning_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M353.07,276.65L395.22,192.35L395.22,156.65L355.22,156.65L355.22,73.07L601.2,73.07Q619.63,73.07 632.21,85.64Q644.78,98.22 644.78,116.65L644.78,153.78L603.35,238.57L475.22,238.57L475.22,201.91L400.48,276.65L353.07,276.65ZM311.87,894.59L311.87,611.5Q311.87,599.07 315.61,586.09Q319.35,573.11 324.59,562.87L475.22,278.57L603.35,278.57Q618.78,294 633.46,319.89Q648.13,345.78 648.13,367.17L648.13,894.59L311.87,894.59Z"/>
-</vector>
