--- feature/habits/src/main/res/drawable/dentistry_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M680,82.61Q747.2,82.61 794.79,130.21Q842.39,177.8 842.39,245Q842.39,256 841.01,274.5Q839.63,293 836.39,317.24L781.39,720.24Q776.15,759.2 746.05,783.79Q715.96,808.39 677,808.39Q653.52,808.39 633.42,798.03Q613.33,787.67 600.09,769.2L493.09,613.43Q491.09,610.15 487.18,608.77Q483.28,607.39 479,607.39Q475.72,607.39 464.91,615.43L360.91,766.43Q346.91,786.67 325.82,797.53Q304.72,808.39 281,808.39Q242.04,808.39 212.45,783.17Q182.85,757.96 177.61,719.24L123.61,317.24Q120.61,293 119.11,274.5Q117.61,256 117.61,245Q117.61,177.8 165.21,130.21Q212.8,82.61 280,82.61Q316.48,82.61 338.34,92.11Q360.2,101.61 380.2,112.85Q399.96,123.85 422.1,133.23Q444.24,142.61 480,142.61Q515.76,142.61 537.9,133.23Q560.04,123.85 579.8,112.85Q599.8,101.61 622.16,92.11Q644.52,82.61 680,82.61Z"/>
-</vector>
