--- feature/habits/src/main/res/drawable/directions_bike_24px.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal"
-    android:autoMirrored="true">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M197.89,888.13Q115.8,888.13 57.9,830.27Q0,772.42 0,690.33Q0,608.24 57.86,550.34Q115.71,492.44 197.8,492.44Q279.89,492.44 337.79,550.29Q395.7,608.15 395.7,690.24Q395.7,772.33 337.84,830.23Q279.98,888.13 197.89,888.13ZM275.2,767.67Q307.09,735.82 307.09,690.33Q307.09,644.83 275.24,612.94Q243.39,581.04 197.89,581.04Q152.39,581.04 120.5,612.89Q88.61,644.74 88.61,690.24Q88.61,735.74 120.46,767.63Q152.31,799.52 197.81,799.52Q243.3,799.52 275.2,767.67ZM435.7,773.63L435.7,571.48L309.37,460.44Q296.65,448.96 290.17,433.69Q283.7,418.42 283.7,401.57Q283.7,384.72 290.43,369.44Q297.17,354.17 308.89,342.46L423.27,228.08Q435.52,215.83 451.41,209.35Q467.29,202.87 484.72,202.87Q502.43,202.87 518.89,209.11Q535.35,215.35 547.83,227.83L625.02,305.02Q652.54,332.54 687.7,348.18Q722.85,363.83 761.91,363.83L761.91,452.44Q704.43,452.44 652.58,430.32Q600.72,408.2 560.24,367.72L529.91,337.39L437.74,429.57L524.3,519.65L524.3,773.63L435.7,773.63ZM566.25,197.34Q542.15,173.24 542.15,139.04Q542.15,104.85 566.25,80.75Q590.35,56.65 624.54,56.65Q658.74,56.65 682.84,80.75Q706.93,104.85 706.93,139.04Q706.93,173.24 682.84,197.34Q658.74,221.43 624.54,221.43Q590.35,221.43 566.25,197.34ZM762.2,888.13Q680.11,888.13 622.21,830.27Q564.3,772.42 564.3,690.33Q564.3,608.24 622.16,550.34Q680.02,492.44 762.11,492.44Q844.2,492.44 902.1,550.29Q960,608.15 960,690.24Q960,772.33 902.14,830.23Q844.29,888.13 762.2,888.13ZM839.5,767.67Q871.39,735.82 871.39,690.33Q871.39,644.83 839.54,612.94Q807.69,581.04 762.19,581.04Q716.7,581.04 684.8,612.89Q652.91,644.74 652.91,690.24Q652.91,735.74 684.76,767.63Q716.61,799.52 762.11,799.52Q807.61,799.52 839.5,767.67Z"/>
-</vector>
