--- feature/habits/src/main/res/drawable/directions_run_24px.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal"
-    android:autoMirrored="true">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M516.41,923.59L516.41,681.91L438.15,607.41L398.87,780.3L115.7,722.87L133.13,635.93L325.13,675.93L387.22,361.5L321.43,386.87L321.43,524.07L234.02,524.07L234.02,329.85L396.57,261.13Q432.28,246.13 449.02,241.39Q465.76,236.65 480,236.65Q502.2,236.65 520.67,248.37Q539.15,260.09 550.87,278.8L591.35,342.8Q616.87,383.85 661.25,410.25Q705.63,436.65 763.59,436.65L763.59,524.07Q696.87,524.07 639.25,497.52Q581.63,470.98 542.15,427.65L520.07,539.28L603.59,618.8L603.59,923.59L516.41,923.59ZM481.23,191.84Q456.65,167.26 456.65,132.59Q456.65,97.91 481.23,73.46Q505.8,49 540.48,49Q575.15,49 599.61,73.46Q624.07,97.91 624.07,132.59Q624.07,167.26 599.61,191.84Q575.15,216.41 540.48,216.41Q505.8,216.41 481.23,191.84Z"/>
-</vector>
