--- feature/habits/src/main/res/drawable/edit_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M111.87,848.13L111.87,662.83L639.39,136.07Q651.87,124.35 667.09,118.11Q682.3,111.87 699.02,111.87Q715.5,111.87 731.22,118.11Q746.93,124.35 758.89,136.83L824.17,202.39Q836.65,214.11 842.51,229.95Q848.37,245.78 848.37,261.74Q848.37,278.46 842.51,293.79Q836.65,309.13 824.17,321.61L297.65,848.13L111.87,848.13ZM698.74,317.5L754.74,261.74L698.5,205.26L642.26,261.26L698.74,317.5Z"/>
-</vector>
