--- feature/habits/src/main/res/drawable/fitness_center_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M532.41,880.07L470.91,818.57L612.91,676.57L283.43,347.09L141.43,489.09L79.93,427.59L135.93,370.07L79.93,314.07L164.41,229.59L108.41,172.07L172.07,108.41L229.59,164.41L314.07,79.93L370.07,135.93L427.59,79.93L489.09,141.43L347.09,283.43L676.57,612.91L818.57,470.91L880.07,532.41L824.07,589.93L880.07,645.93L795.59,730.41L851.59,787.93L787.93,851.59L730.41,795.59L645.93,880.07L589.93,824.07L532.41,880.07Z"/>
-</vector>
