--- feature/habits/src/main/res/drawable/language_chinese_dayi_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M706.3,622.17Q732.15,622.17 757.82,610.01Q783.48,597.85 805.72,575.52L805.72,475.02Q783.91,478.02 765.25,482.02Q746.59,486.02 730.07,491.02Q687.22,504.54 666.27,523.99Q645.33,543.43 645.33,569.52Q645.33,593.85 662.01,608.01Q678.7,622.17 706.3,622.17ZM681.87,695.43Q625.35,695.43 592.23,662.93Q559.11,630.43 559.11,574.43Q559.11,521.96 592.59,488.36Q626.07,454.76 699.07,434.52Q721.35,428.52 747.53,423.76Q773.72,419 805.48,415L805.48,415Q803.48,370.39 784.67,349.97Q765.87,329.54 726.5,329.54Q700.74,329.54 674.64,339.4Q648.54,349.26 609.02,373.74L574.15,312.48Q608.83,286.04 654.16,270.18Q699.5,254.33 745.26,254.33Q816.98,254.33 854.46,299.16Q891.93,344 891.93,429.43L891.93,686.43L819.2,686.43L813.2,643.11Q785.2,667.63 751.93,681.53Q718.67,695.43 681.87,695.43ZM99.74,707.39Q88.02,691.96 77.75,681.64Q67.48,671.33 56.61,664.89Q129.5,633.57 175.5,582.04Q221.5,530.52 242.93,444.83L76.83,444.83L76.83,389.93L254.52,389.93Q259,359.7 261.24,325.7Q263.48,291.7 263.48,254.17L320.65,254.17Q320.65,291.7 318.79,325.2Q316.93,358.7 311.93,389.93L505.59,389.93L505.59,444.83L334.48,444.83Q361.2,522.24 408.52,577.59Q455.85,632.93 518.22,659.17Q508.78,668.85 500.09,679.64Q491.39,690.43 479.91,707.39Q417.87,679.15 369.47,622.37Q321.07,565.59 290.35,485.26Q265.15,566.11 218.01,620.11Q170.87,674.11 99.74,707.39Z"/>
-</vector>
