--- feature/habits/src/main/res/drawable/local_laundry_service_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M242.87,888.13Q205.09,888.13 178.48,861.52Q151.87,834.91 151.87,797.13L151.87,162.87Q151.87,125.09 178.48,98.48Q205.09,71.87 242.87,71.87L717.13,71.87Q754.91,71.87 781.52,98.48Q808.13,125.09 808.13,162.87L808.13,797.13Q808.13,834.91 781.52,861.52Q754.91,888.13 717.13,888.13L242.87,888.13ZM619.47,699.47Q677.13,641.8 677.13,560Q677.13,478.2 619.47,420.53Q561.8,362.87 480,362.87Q398.2,362.87 340.53,420.53Q282.87,478.2 282.87,560Q282.87,641.8 340.53,699.47Q398.2,757.13 480,757.13Q561.8,757.13 619.47,699.47ZM480,686.98Q454.72,686.98 431.29,677.84Q407.87,668.7 389.59,650.41L570.41,469.59Q588.7,487.87 597.84,511.29Q606.98,534.72 606.98,560Q606.98,612.85 569.91,649.91Q532.85,686.98 480,686.98ZM325.02,287.41Q342.98,287.41 355.2,275.2Q367.41,262.98 367.41,245.02Q367.41,227.3 355.08,215.09Q342.74,202.87 325.02,202.87Q307.3,202.87 295.09,215.09Q282.87,227.3 282.87,245.02Q282.87,262.74 295.09,275.08Q307.3,287.41 325.02,287.41ZM479.5,275.08Q491.72,262.74 491.72,245.02Q491.72,227.3 479.5,215.09Q467.28,202.87 449.57,202.87Q431.85,202.87 419.63,215.09Q407.41,227.3 407.41,245.02Q407.41,262.74 419.63,275.08Q431.85,287.41 449.57,287.41Q467.28,287.41 479.5,275.08Z"/>
-</vector>
