--- feature/habits/src/main/res/drawable/nutrition_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M278.15,761.85Q195.22,678.91 195.22,560Q195.22,466.72 249.28,392.1Q303.35,317.48 391.67,289.33Q358.2,280.61 333.58,259.03Q308.96,237.46 294.2,207.74Q279.43,178.02 273.79,143.07Q268.15,108.11 272.39,72.39Q314.35,66.91 352.66,76.67Q390.98,86.43 422.17,108.67Q453.37,130.91 473.71,162.51Q494.04,194.11 498.28,234.07Q511.52,202.83 529.78,175.09Q548.04,147.35 571.28,124.11Q583.72,111.67 602.39,111.67Q621.07,111.67 633.5,124.11Q645.93,136.54 645.93,155.22Q645.93,173.89 633.5,186.33Q611.74,208.09 595.34,234.23Q578.93,260.37 570.7,290.13Q658.22,319.57 711.5,393.66Q764.78,467.76 764.78,560Q764.78,678.91 681.85,761.85Q598.91,844.78 480,844.78Q361.09,844.78 278.15,761.85Z"/>
-</vector>
