--- feature/habits/src/main/res/drawable/pets_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M173.3,481.41Q130.6,481.41 101.12,451.82Q71.63,422.22 71.63,379.5Q71.63,336.78 101.12,307.18Q130.61,277.59 173.31,277.59Q216.02,277.59 245.62,307.18Q275.22,336.78 275.22,379.5Q275.22,422.22 245.66,451.82Q216.11,481.41 173.3,481.41ZM285.18,288.82Q255.7,259.33 255.7,216.62Q255.7,173.91 285.18,144.32Q314.67,114.72 357.38,114.72Q400.09,114.72 429.68,144.27Q459.28,173.83 459.28,216.63Q459.28,259.33 429.68,288.82Q400.09,318.3 357.38,318.3Q314.67,318.3 285.18,288.82ZM530.32,288.82Q500.72,259.33 500.72,216.62Q500.72,173.91 530.27,144.32Q559.83,114.72 602.63,114.72Q645.33,114.72 674.82,144.32Q704.3,173.91 704.3,216.62Q704.3,259.33 674.82,288.82Q645.33,318.3 602.62,318.3Q559.91,318.3 530.32,288.82ZM786.69,481.41Q743.98,481.41 714.38,451.82Q684.78,422.22 684.78,379.5Q684.78,336.78 714.34,307.18Q743.89,277.59 786.7,277.59Q829.4,277.59 858.88,307.18Q888.37,336.78 888.37,379.5Q888.37,422.22 858.88,451.82Q829.39,481.41 786.69,481.41ZM263.85,890.02Q217.41,890.02 186.2,854.68Q154.98,819.35 154.98,770.91Q154.98,717.96 191.2,678.6Q227.41,639.24 262.41,600.28Q291.17,568.33 312.65,531.71Q334.13,495.09 363.37,462.85Q385.83,436.23 415.44,418.82Q445.04,401.41 480,401.41Q514.96,401.41 544.55,417.89Q574.15,434.37 596.87,460.85Q625.15,493.14 646.87,530.47Q668.59,567.8 697.59,600.28Q732.59,639.24 768.8,678.6Q805.02,717.96 805.02,770.91Q805.02,819.35 773.8,854.68Q742.59,890.02 696.15,890.02Q641.43,890.02 588.08,880.9Q534.72,871.78 480,871.78Q425.28,871.78 371.92,880.9Q318.57,890.02 263.85,890.02Z"/>
-</vector>
