--- feature/habits/src/main/res/drawable/recycling_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M366.09,365.61L456.52,215.98L396.57,116.07Q384.09,96.07 361.35,96.07Q338.61,96.07 326.37,116.07L226.46,282.17L366.09,365.61ZM760.02,642.63L669.59,491.76L810.98,410.33L876.17,519.48Q887.17,536.96 888.17,558.2Q889.17,579.43 879.17,597.91Q868.93,618.39 849.2,630.51Q829.46,642.63 805.98,642.63L760.02,642.63ZM642.87,927.65L480,764.78L642.87,601.91L642.87,683.35L836.46,683.35L777.26,801.5Q766.26,821.98 746.78,834.1Q727.3,846.22 704.07,846.22L642.87,846.22L642.87,927.65ZM248.7,846.22Q228.46,846.22 211.72,835.6Q194.98,824.98 186.74,807.5Q178.74,791.26 179.12,773.52Q179.5,755.78 188.74,740.3L223.46,683.35L398.57,683.35L398.57,846.22L248.7,846.22ZM148.02,730.3L81.83,597.91Q72.35,579.43 73.09,558.7Q73.83,537.96 84.83,519.48L101.3,491.76L31.87,450.04L254.93,394.09L310.89,618.15L240.7,575.2L148.02,730.3ZM697.83,381.85L475,325.89L545.2,284.17L417.8,72.35L561.43,72.35Q582.91,72.35 601.53,83.09Q620.15,93.83 631.63,112.07L684.59,200.74L754.02,157.78L697.83,381.85Z"/>
-</vector>
