--- feature/habits/src/main/res/drawable/remove_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M191.87,525.5L191.87,434.5L768.13,434.5L768.13,525.5L191.87,525.5Z"/>
-</vector>
