--- feature/habits/src/main/res/drawable/self_improvement_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M272,803.59Q240.33,803.59 218.37,781.63Q196.41,759.67 196.41,728Q196.41,705.8 209.13,686.35Q221.85,666.89 242.57,658.65L396.41,597.37L396.41,515.5Q342.41,576.11 271.15,608.05Q199.89,640 116.41,640L116.41,552.59Q185.37,552.59 241.23,525.07Q297.09,497.54 341.13,446.02L395.13,381.78Q407.61,367.07 424.33,359.83Q441.04,352.59 460,352.59L500,352.59Q518.96,352.59 535.67,359.95Q552.39,367.3 564.87,382.02L618.87,446.5Q662.91,498.02 718.77,525.3Q774.63,552.59 843.59,552.59L843.59,640Q760.11,640 688.85,608.05Q617.59,576.11 563.59,515.5L563.59,597.37L717.43,658.65Q738.15,666.89 750.87,686.35Q763.59,705.8 763.59,728Q763.59,759.67 741.63,781.63Q719.67,803.59 688,803.59L400,803.59L400,780Q400,753.28 417.36,736.64Q434.72,720 460,720L580,720Q589,720 594.5,714.5Q600,709 600,700Q600,691 594.5,685.5Q589,680 580,680L460,680Q418.24,680 389.12,709Q360,738 360,780L360,803.59L272,803.59ZM420.87,288.13Q396.41,263.67 396.41,229Q396.41,194.33 420.87,169.87Q445.33,145.41 480,145.41Q514.67,145.41 539.13,169.87Q563.59,194.33 563.59,229Q563.59,263.67 539.13,288.13Q514.67,312.59 480,312.59Q445.33,312.59 420.87,288.13Z"/>
-</vector>
