--- feature/habits/src/main/res/drawable/shopping_basket_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M226.98,848.13Q196.39,848.13 172.16,829.48Q147.93,810.83 139.22,781.24L36.26,409.8Q30.54,388.17 43.72,370.02Q56.89,351.87 79.76,351.87L267.13,351.87L439.3,93.7Q445.5,83.78 456.29,77.83Q467.09,71.87 479,71.87Q490.91,71.87 501.71,77.83Q512.5,83.78 518.7,93.7L690.87,351.87L880.24,351.87Q903.11,351.87 916.28,370.02Q929.46,388.17 923.74,409.8L821.5,781Q812.78,810.83 788.67,829.48Q764.57,848.13 733.74,848.13L226.98,848.13ZM537.93,657.93Q561.91,633.96 561.91,600Q561.91,566.04 537.93,542.07Q513.96,518.09 480,518.09Q446.04,518.09 422.07,542.07Q398.09,566.04 398.09,600Q398.09,633.96 422.07,657.93Q446.04,681.91 480,681.91Q513.96,681.91 537.93,657.93ZM376.09,351.87L582.91,351.87L478.76,198.22L376.09,351.87Z"/>
-</vector>
