--- feature/habits/src/main/res/drawable/shower_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M316.41,729.09Q297.98,729.09 285.28,716.56Q272.59,704.02 272.59,685.5Q272.59,666.98 285.28,654.44Q297.98,641.91 316.41,641.91Q334.85,641.91 347.42,654.44Q360,666.98 360,685.5Q360,704.02 347.47,716.56Q334.94,729.09 316.41,729.09ZM480,729.09Q461.48,729.09 448.94,716.56Q436.41,704.02 436.41,685.5Q436.41,666.98 448.94,654.44Q461.48,641.91 480,641.91Q498.52,641.91 511.06,654.44Q523.59,666.98 523.59,685.5Q523.59,704.02 511.06,716.56Q498.52,729.09 480,729.09ZM643.59,729.09Q625.15,729.09 612.58,716.56Q600,704.02 600,685.5Q600,666.98 612.53,654.44Q625.06,641.91 643.59,641.91Q662.02,641.91 674.72,654.44Q687.41,666.98 687.41,685.5Q687.41,704.02 674.72,716.56Q662.02,729.09 643.59,729.09ZM191.87,561.91L191.87,470.91Q191.87,364.2 260.95,285Q330.02,205.8 434.5,190.37L434.5,103.5L525.5,103.5L525.5,190.37Q629.98,205.8 699.05,284.88Q768.13,363.96 768.13,470.91L768.13,561.91L191.87,561.91ZM316.41,856.5Q297.98,856.5 285.28,843.97Q272.59,831.44 272.59,812.91Q272.59,794.48 285.28,781.78Q297.98,769.09 316.41,769.09Q334.85,769.09 347.42,781.78Q360,794.48 360,812.91Q360,831.35 347.47,843.92Q334.94,856.5 316.41,856.5ZM480,856.5Q461.48,856.5 448.94,843.97Q436.41,831.44 436.41,812.91Q436.41,794.48 448.94,781.78Q461.48,769.09 480,769.09Q498.52,769.09 511.06,781.78Q523.59,794.48 523.59,812.91Q523.59,831.35 511.06,843.92Q498.52,856.5 480,856.5ZM643.59,856.5Q625.15,856.5 612.58,843.97Q600,831.44 600,812.91Q600,794.48 612.53,781.78Q625.06,769.09 643.59,769.09Q662.02,769.09 674.72,781.78Q687.41,794.48 687.41,812.91Q687.41,831.35 674.72,843.92Q662.02,856.5 643.59,856.5Z"/>
-</vector>
