--- feature/habits/src/main/res/drawable/washoku_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M275.7,848.13L213.74,661.02Q177.26,640.07 152.14,606.23Q127.02,572.39 115.78,529.91L82.11,398.8L878.13,398.8L844.46,529.91Q833.22,572.39 808.1,606.23Q782.98,640.07 746.5,661.02L684.54,848.13L275.7,848.13ZM806.46,318.8L153.78,318.8L153.78,318.8Q153.78,266.17 189.84,229.38Q225.89,192.59 277.04,192.59Q284.33,192.59 292.49,193.85Q300.65,195.11 308.65,196.87Q326.61,158.63 361.83,135.25Q397.04,111.87 439.76,111.87Q487.28,111.87 524.5,139.53Q561.72,167.2 576.63,212.24Q595.87,193.43 620.39,182.89Q644.91,172.35 671.96,172.35Q728.11,172.35 767.28,212.14Q806.46,251.93 806.46,308.57L806.46,318.8Z"/>
-</vector>
