--- feature/habits/src/main/res/drawable/water_drop_24px.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960"
-    android:tint="?attr/colorControlNormal">
-  <path
-      android:fillColor="@android:color/white"
-      android:pathData="M491.72,758.8Q504.91,757.8 514.13,748.59Q523.35,739.37 523.35,726.41Q523.35,711.7 513.51,702.36Q503.67,693.02 488.72,694.02Q446.28,697.02 403.39,671.76Q360.5,646.5 348.74,583.33Q346.74,571.85 337.76,564.13Q328.78,556.41 316.35,556.41Q301.15,556.41 291.43,567.63Q281.72,578.85 284.96,593.8Q301.96,685.04 364.84,723.92Q427.72,762.8 491.72,758.8ZM245.64,791.86Q151.87,695.59 151.87,552Q151.87,450.09 233.28,330.08Q314.7,210.07 480,69.48Q645.3,210.07 726.72,330.08Q808.13,450.09 808.13,552Q808.13,695.59 714.36,791.86Q620.59,888.13 480,888.13Q339.41,888.13 245.64,791.86Z"/>
-</vector>
